Fix expiration time check
This useEffect checks if the session has expired and creates a new one
if that's the case. But here we were comparing a string containing the
UNIX timestamp with a string containing a date in human-readable format.

This caused that the session was not being renewed when it was expired.