Fix 2h TACo auth expiration time on SIWE
The maximum expiration time is not defined for SIWE standard, so it can
be months (like in this case).

However, for TACo this expiration date is fixed to two hours after
issuing the certificate.

So, when loading a SIWE session, this can be one that is still valid for
SIWE standard, but not for TACo (more than 2h since it was issued).